# Apple_SDE_Submission/submission/api_openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Pricing Pipeline API",
        "description": "Multi-partner pricing ingestion + harmonisation + Data Quality + anomaly detection.\n\n**Path A \u2014 POST /pipeline**: one-click 9-step orchestrator.\n**Path B \u2014 sub-modules**: /load-data, /compute-dq, /detect-anomalies, /harmonise-product. Independently callable; combine in sequence to cover the same 9 steps with degraded gating semantics.",
        "contact": {
            "name": "huizhongwu"
        },
        "version": "0.2.0"
    },
    "paths": {
...
        "/pipeline": {
            "post": {
                "tags": [
                    "pipeline"
                ],
                "summary": "One-click orchestrator \u2014 runs all 9 steps end-to-end",
                "description": "Runs the canonical 9-step pipeline in interleaved order:\n\n  1. Parse CSV \u2192 `stg_price_offer`\n  2. INGEST-stage Data Quality\n  3. Harmonise raw product names\n  4. PRE_FACT-stage Data Quality (HIGH-severity gate)\n  5. Insert into `fact_price_offer` + payment child (**only PRE_FACT-passing rows**)\n  6. SEMANTIC-stage Data Quality (soft signals; flag-and-keep)\n  7. Update Slowly Changing Dimension Type 2 history\n  8. Detect pricing anomalies on the new batch\n  9. Write per-batch summary to `dws_partner_dq_per_batch`\n\nSingle transaction. PRE_FACT bad rows are hard-blocked from `fact_price_offer`. Equivalent to calling /load-data + /compute-dq + /detect-anomalies in sequence except for stronger gating semantics (the sub-module path lets bad rows enter fact and flags them post-hoc).",
                "operationId": "pipeline_pipeline_post",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/Body_pipeline_pipeline_post"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PipelineResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/load-data": {
            "post": {
                "tags": [
                    "load-data"
                ],
                "summary": "Load CSV into the table (parse + harmonise + write fact + Slowly Changing Dimension Type 2)",
                "description": "Task B sub-module \u2014 covers pipeline steps **1, 3, 5, 7, 9**:\n\n  1. Parse CSV into `stg_price_offer`\n  3. Harmonise raw product names against the canonical registry\n  5. Insert change events into `fact_price_offer` + payment child (**no PRE_FACT gate** \u2014 all rows with parseable fields enter fact; Data Quality flagging happens post-hoc via /compute-dq)\n  7. Update Slowly Changing Dimension Type 2 history\n  9. Write batch summary\n\nSteps 2/4/6 (Data Quality) and step 8 (anomaly detection) are NOT run here \u2014 call /compute-dq and /detect-anomalies separately, or use POST /pipeline for end-to-end with hard gating.",
                "operationId": "load_data_load_data_post",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/Body_load_data_load_data_post"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "202": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/LoadDataAcceptedResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/load-data/{job_id}": {
            "get": {
                "tags": [
                    "load-data"
                ],
                "summary": "Poll the progress / status of a load job",
                "operationId": "get_load_job_load_data__job_id__get",
                "parameters": [
                    {
                        "name": "job_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "description": "The job_id returned by POST /load-data",
                            "title": "Job Id"
                        },
                        "description": "The job_id returned by POST /load-data"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/LoadJobStatus"
                                }
                            }
                        }
                    },
                    "404": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        },
                        "description": "Not Found"
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/compute-dq": {
            "post": {
                "tags": [
                    "compute-dq"
                ],
                "summary": "Validate Data Quality \u2014 runs all 13 rules and writes to dq_output + dq_bad_records",
                "description": "Task B sub-module \u2014 covers pipeline steps **2, 4, 6, 9**:\n\n  2. INGEST-stage Data Quality (parse / format / required-field rules)\n  4. PRE_FACT-stage Data Quality (country\u2194currency, partner\u2194country, harmonise match)\n  6. SEMANTIC-stage Data Quality (soft signals \u2014 low-confidence harmonise, category sanity bounds)\n  9. Refresh batch summary\n\nAll 13 rules execute against the batch. Per-rule pass rates are stored in **`dq_output`**; per-row violations (with the original CSV payload preserved as JSONB) are stored in **`dq_bad_records`**.\n\nDesigned to be called AFTER /load-data has populated fact_price_offer. PRE_FACT-failing rows are flagged post-hoc \u2014 they remain in fact and must be filtered via `LEFT JOIN dq_bad_records` if downstream queries need a clean view. For end-to-end execution with hard gating, use POST /pipeline instead.",
                "operationId": "compute_dq_compute_dq_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ComputeDQRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ComputeDQResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/detect-anomalies": {
            "post": {
                "tags": [
                    "detect-anomalies"
                ],
                "summary": "Detect pricing anomalies + return visualization payload",
                "description": "Task B sub-module \u2014 covers pipeline step **8** (and refreshes step 9 summary).\n\nFor each offer in scope, compares its USD price against the 30-day rolling mean from `fact_partner_price_history` (the Slowly Changing Dimension Type 2 table). Returns the anomaly list with severity classification (HIGH \u226525%, MEDIUM \u226515%, LOW \u226510%) and a structured visualization payload (time series + baseline band + cross-partner comparison) that the frontend can render directly via Chart.js / Recharts.",
                "operationId": "detect_anomalies_detect_anomalies_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/DetectAnomaliesRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DetectAnomaliesResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/harmonise-product": {
            "get": {
                "tags": [
                    "harmonise-product"
                ],
                "summary": "Harmonise a raw product name to canonical model(s) \u2014 Top-K + score",
                "description": "Task B sub-module \u2014 exposes the harmoniser as an ad-hoc lookup. Returns up to **k** ranked candidate matches for a partner-supplied product name. Each candidate has a 0-1 score, a confidence bucket (HIGH / MEDIUM / LOW / MANUAL) and a per-signal breakdown for explainability.\n\nInternally, pipeline step 3 (`harmonise_in_stg`) calls this same harmoniser for every staged row during /load-data and /pipeline.",
                "operationId": "harmonise_product_harmonise_product_get",
                "parameters": [
                    {
                        "name": "q",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "minLength": 1,
                            "description": "Raw product name to harmonise",
                            "title": "Q"
                        },
                        "description": "Raw product name to harmonise"
                    },
                    {
                        "name": "k",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "maximum": 20,
                            "minimum": 1,
                            "description": "Number of candidates to return",
                            "default": 5,
                            "title": "K"
                        },
                        "description": "Number of candidates to return"
                    },
                    {
                        "name": "min_confidence",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "$ref": "#/components/schemas/Confidence",
                            "description": "Minimum confidence bucket",
                            "default": "LOW"
                        },
                        "description": "Minimum confidence bucket"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HarmoniseResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/bad-records": {
            "get": {
                "tags": [
                    "business-review"
                ],
                "summary": "List records flagged by Data Quality rules",
                "description": "Powers the business-user review UI. Default filter: `status=NEW` to surface unresolved items. Pagination via `page` / `page_size`.",
                "operationId": "list_bad_records_bad_records_get",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "$ref": "#/components/schemas/BadRecordStatus",
                            "default": "NEW"
                        }
                    },
                    {
                        "name": "severity",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "$ref": "#/components/schemas/Severity"
                        }
                    },
                    {
                        "name": "assignee",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "title": "Assignee"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "minimum": 1,
                            "default": 1,
                            "title": "Page"
                        }
                    },
                    {
                        "name": "page_size",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "maximum": 500,
                            "minimum": 1,
                            "default": 50,
                            "title": "Page Size"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/BadRecordList"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Body_pipeline_pipeline_post": {
                "properties": {
                    "file": {
                        "type": "string",
                        "format": "binary",
                        "title": "File",
                        "description": "CSV file"
                    },
                    "partner_code": {
                        "type": "string",
                        "title": "Partner Code",
                        "description": "Partner identifier registered in dim_partner"
                    }
                },
                "type": "object",
                "required": [
                    "file",
                    "partner_code"
                ],
                "title": "Body_pipeline_pipeline_post"
            },
...
            "Confidence": {
                "type": "string",
                "enum": [
                    "HIGH",
                    "MEDIUM",
                    "LOW",
                    "MANUAL"
                ],
                "title": "Confidence"
            },
...
            "ErrorResponse": {
                "properties": {
                    "error": {
                        "type": "string",
                        "title": "Error"
                    },
                    "detail": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Detail"
                    },
                    "code": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Code"
                    }
                },
                "type": "object",
                "required": [
                    "error"
                ],
                "title": "ErrorResponse"
            },
...
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
...
        }
    }
}